+++ TwoWheelTrader/Models/Enduro/Honda.cs
@@ -4,7 +4,7 @@ namespace TwoWheelTrader.Models.Enduro
 {
     public class Honda : IEnduro
     {
-        public Honda(string model, int cc, int year, int priceForeign, string link, int distance)
+        public Honda(string model, int cc, int year, double priceForeign, string link, int distance)
         {
             Make = "Honda";
             Model = model;
