+++ TwoWheelTrader/Models/Enduro/KTM.cs
@@ -4,7 +4,7 @@ namespace TwoWheelTrader.Models.Enduro
 {
     public class KTM : IEnduro
     {
-        public KTM(string model, int cc, int year, int priceForeign, string link, int distance)
+        public KTM(string model, int cc, int year, double priceForeign, string link, int distance)
         {
             Make = "KTM";
             Model = model;
